move globals.ts from build to root
