+++ demo/src/actions.ts
@@ -1,4 +1,4 @@
-import '@grafana/agent-web/build/cjs/globals';
+import '@grafana/agent-web/globals';
 import { SpanStatusCode } from '@opentelemetry/api';
 
 const w = window as any;
