+++ packages/web/globals.ts
