+++ packages/web/package.json
@@ -16,6 +16,7 @@
   "types": "dist/cjs/index.d.ts",
   "files": [
     "dist",
+    "globals.ts",
     "README.md",
     "LICENSE"
   ],
